bugfix rule ImplementePrivateEndpointApis (#463)
* fix rule ImplementePrivateEndpointApis

* update test

* fix typo

* fix test

* changelog

---------
