+++ common/changes/@microsoft.azure/openapi-validator-core/bugfixes_2023-02-21-10-19.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@microsoft.azure/openapi-validator-core",
+      "comment": "bugfix for dependency detection",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@microsoft.azure/openapi-validator-core"
+}
