+++ common/changes/@microsoft.azure/openapi-validator-rulesets/bugfixes_2023-02-21-10-19.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@microsoft.azure/openapi-validator-rulesets",
+      "comment": "re-implement rule ImplementPrivateEndpointApis",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@microsoft.azure/openapi-validator-rulesets"
+}
