+++ packages/rulesets/src/native/rulesets/common.ts
@@ -40,7 +40,6 @@ require("../legacyRules/UniqueModelName")
 require("../legacyRules/MissingXmsErrorResponse")
 require("../legacyRules/MissingTypeObject")
 require("../legacyRules/PrivateEndpointResourceSchemaValidation")
-require("../legacyRules/ImplementPrivateEndpointAPIs")
 require("../legacyRules/ParametersOrder")
 require("../legacyRules/ExtensionResourcePathPattern")
 require("../legacyRules/XmsEnumValidation")
